+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `UseOffsetPagingAttribute` takes optional arguments for default page size and max page size
 - `EntityGraphQL.AspNet` now uses System.Text.Json for de/serialization internally. It still supports JSON.NET objects in the variable object (as well as System.Text.Json types)
 - Fix #19 - you can mark fields (including mutations) as deprecated using the dotnet `[Obsolete]` attribute or the `IField.Deprecate(reason)` method in the schema building
+- Fix #88 - If an argument is marked as required and not provided in the query request an error is raised. Mark arguments optional via the API, or make it a nullable type or give it a default value
 
 ## Obsolete - will be removed 2.x
 - `RequireAnyClaim()` & `RequireAllClaims()`. Use `RequireAnyRole()` & `RequireAllRoles()` as the `ClaimTypes.Role` was used previously and this change makes it explicit